requirements.txt
Flask
firebase-admin
scikit-learn
numpy
colorama
requests
python-dotenv
